stores: extend the float-constant store run to `double` globals (more BYTE coverage)
Follow-up to 8594194. `double` global constant runs (`gd=1.0; ge=2.0;`) use the
IDENTICAL pre-load-distinct-FPR schedule as the float case, just `lfd`/`stfd` off an
8-byte pool: `lfd f1,@a; lfd f0,@b; stfd f1,gd; stfd f0,ge`. They were still a DIFF.

Generalize try_float_constant_store_fill to accept a HOMOGENEOUS run of `float` OR
`double` global stores (a mixed float/double run defers): pick lfs/lfd via
load_float_literal(_, _, run_is_double), and key prematerialized_float_constants on
the literal's full f64 bits (widened u32->u64) so the double values match correctly.

Verified BYTE: 2/3 distinct `double` runs + a same-value double run; float runs
unchanged (still BYTE); a mixed float/double run defers (declined, not miscompiled).
Gate GREEN (vreg 56/0, oracle 1163/0 x5, support 454B/0D/0DEFER). (Pre-existing,
unrelated: a SINGLE double global store still DIFFs — a separate latent issue this
run handler does not touch.)

+++ crates/pipeline/mwcc-syntax-trees-to-machine-code/src/generator.rs
@@ -195,11 +195,12 @@ pub(crate) struct Generator {
     /// the wrong (interleaved, r0-reusing) order. Constant-index stores use a
     /// displacement (no r0 scaling) and never set or consult this.
     pub(crate) emitted_variable_index_store: bool,
-    /// Float constants pre-loaded into fixed FPRs for a distinct-float-constant store run
-    /// (`gf=1.0f; gg=2.0f`): mwcc pre-loads each into a distinct FPR (`lfs f1,@a; lfs f0,@b;
-    /// stfs f1,gf; stfs f0,gg`), so `place_store_value` reuses the pre-loaded FPR by float
-    /// bits instead of re-pooling/re-loading. `(f32 bits, FPR)`; empty outside a run.
-    pub(crate) prematerialized_float_constants: Vec<(u32, u8)>,
+    /// Float/double constants pre-loaded into fixed FPRs for a distinct-float-constant store
+    /// run (`gf=1.0f; gg=2.0f`, or the `double` `lfd` variant): mwcc pre-loads each into a
+    /// distinct FPR (`lfs f1,@a; lfs f0,@b; stfs f1,gf; stfs f0,gg`), so `place_store_value`
+    /// reuses the pre-loaded FPR by the literal's f64 bits instead of re-pooling/re-loading.
+    /// `(FloatLiteral f64 bits, FPR)`; empty outside a run (runs are homogeneous float/double).
+    pub(crate) prematerialized_float_constants: Vec<(u64, u8)>,
     /// Address-taken variables and their stack-frame slots. A name here is
     /// frame-resident: `&v` and type-punned accesses read/write its slot.
     pub(crate) frame_slots: HashMap<String, FrameSlot>,
